Update regression.py
The previous change to this file causes regression testing on Windows under a regular command prompt to bomb out - git complains that it doesn't understand the option "--hard;".  This is a command line issue - you can't separate commands with a ";" under windows git (as far as I know).

Suggested fix is to break this into 2 commands.